+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+
+- Fixes an issue with improper permissions when creating a game embed.
+
 ## [v11.2.1](https://github.com/lexicalunit/spellbot/releases/tag/v11.2.1) - 2024-06-24
 
 ### Changed